requirements.txt
# Core AI
openai>=1.30.0
langgraph>=0.1.0

# Data validation
pydantic>=2.0.0

# API
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Utilities
python-dotenv>=1.0.0
fastmcp>=2.0.0
PyYAML>=6.0.0

# Testing
pytest>=8.0.0
